backend/requirements.txt
Flask==3.1.0
flask_cors==5.0.1
mysql_connector_repackaged==0.3.1
opencv_python==4.8.1.78
opencv_python==4.11.0.86
ultralytics==8.3.82
ultralytics==8.3.94
